1) balance trigger slash badges. 2) added new sound for runner jump. 3) Fix bug: FloatyUtil throwing invalid cast exception

+++ MyZooPets/Assets/Scripts/TriggerNinja/NinjaManager.cs
@@ -397,7 +397,7 @@ public class NinjaManager : MinigameManager<NinjaManager> {
 	        option.Add("text", strText);
 			option.Add("prefab", "NinjaComboFloatyText");
 			option.Add("position", vPos);
-	        option.Add("textSize", Constants.GetConstant<int>("Ninja_ComboTextSize"));
+	        option.Add("textSize", Constants.GetConstant<float>("Ninja_ComboTextSize"));
 			
 			// spawn floaty text
 	        FloatyUtil.SpawnFloatyText(option);
